app/requirements.txt
fastapi
uvicorn
pandas
numpy
matplotlib
scikit-learn
google-cloud-bigquery
google-cloud-storage
python-dotenv
